backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aiomysql
pymysql
alembic
pydantic
pydantic-settings
qdrant-client
insightface
onnxruntime
celery
redis
python-multipart
python-jose[cryptography]
bcrypt
beautifulsoup4
opencv-python-headless
numpy
Pillow
pytest
httpx
